fix(folio): parse header footer text boxes
