+++ packages/folio/src/core/docx/blockContentParser.ts
@@ -1,0 +1,426 @@
+/**
+ * Shared OOXML block-content parser.
+ *
+ * The document body, headers, footers, and SDT content all expose the same
+ * block-level model: paragraphs, tables, and nested structured document tags.
+ * Keeping the parser shared prevents body-only fixes, especially for drawings
+ * like text boxes that can appear in headers and footers too.
+ */
+
+import type {
+  MediaFile,
+  Paragraph,
+  RelationshipMap,
+  Run,
+  Shape,
+  ShapeContent,
+  Table,
+  Theme,
+} from "../types/document";
+import { convertBulletToUnicode } from "./bulletMarkers";
+import type { NumberingMap } from "./numberingParser";
+import { parseParagraph } from "./paragraphParser";
+import type { StyleMap } from "./styleParser";
+import { parseTable } from "./tableParser";
+import {
+  getTextBoxContentElement,
+  isTextBoxDrawing,
+  parseTextBox,
+  parseTextBoxContent,
+} from "./textBoxParser";
+import { findDeep, getChildElements, getLocalName } from "./xmlParser";
+import type { XmlElement } from "./xmlParser";
+
+type ParseBlockContentOptions = {
+  inHeaderFooter?: boolean;
+};
+
+const toRoman = (numParam: number): string => {
+  let num = numParam;
+  const romanNumerals: [number, string][] = [
+    [1000, "M"],
+    [900, "CM"],
+    [500, "D"],
+    [400, "CD"],
+    [100, "C"],
+    [90, "XC"],
+    [50, "L"],
+    [40, "XL"],
+    [10, "X"],
+    [9, "IX"],
+    [5, "V"],
+    [4, "IV"],
+    [1, "I"],
+  ];
+
+  let result = "";
+  for (const [value, symbol] of romanNumerals) {
+    while (num >= value) {
+      result += symbol;
+      num -= value;
+    }
+  }
+  return result;
+};
+
+const formatNumber = (value: number, numFmt: string): string => {
+  switch (numFmt) {
+    case "decimal":
+    case "decimalZero":
+      return String(value);
+    case "lowerLetter":
+      return String.fromCodePoint(96 + ((value - 1) % 26) + 1);
+    case "upperLetter":
+      return String.fromCodePoint(64 + ((value - 1) % 26) + 1);
+    case "lowerRoman":
+      return toRoman(value).toLowerCase();
+    case "upperRoman":
+      return toRoman(value);
+    case "bullet":
+      return "\u2022";
+    default:
+      return String(value);
+  }
+};
+
+const computeListMarker = (
+  paragraph: Paragraph,
+  numbering: NumberingMap | null,
+  listCounters: Map<number, number[]>,
+  abstractCounters: Map<number, number[]>,
+): void => {
+  const listRendering = paragraph.listRendering;
+  if (!listRendering || !numbering) {
+    return;
+  }
+
+  const { numId, level } = listRendering;
+  if (numId === 0) {
+    return;
+  }
+
+  if (!listCounters.has(numId)) {
+    listCounters.set(numId, Array.from<number>({ length: 9 }).fill(0));
+  }
+
+  const counters = listCounters.get(numId);
+  if (!counters) {
+    return;
+  }
+
+  const abstractNumId = numbering.getAbstractNumId(numId);
+  if (abstractNumId !== null && level > 0) {
+    const latestAbstractCounters = abstractCounters.get(abstractNumId);
+    const missingParentCounters = counters
+      .slice(0, level)
+      .every((value) => value === 0);
+    if (latestAbstractCounters && missingParentCounters) {
+      for (let i = 0; i < level; i += 1) {
+        counters[i] = latestAbstractCounters[i] ?? 0;
+      }
+    }
+  }
+
+  counters[level] = (counters[level] || 0) + 1;
+
+  for (let i = level + 1; i < counters.length; i += 1) {
+    counters[i] = 0;
+  }
+
+  if (abstractNumId !== null) {
+    abstractCounters.set(abstractNumId, [...counters]);
+  }
+
+  const pattern = listRendering.marker;
+
+  if (listRendering.isBullet) {
+    listRendering.marker = convertBulletToUnicode(pattern || "");
+    return;
+  }
+
+  let computedMarker = pattern;
+  const currentLevelInfo = numbering.getLevel(numId, level);
+  const useLegalNumbering =
+    currentLevelInfo?.isLgl === true || listRendering.isLegal === true;
+
+  for (let lvl = 0; lvl <= level; lvl += 1) {
+    const placeholder = `%${lvl + 1}`;
+    if (computedMarker.includes(placeholder)) {
+      const value = counters[lvl] ?? 0;
+      const levelInfo = numbering.getLevel(numId, lvl);
+      const formatted = formatNumber(
+        value,
+        useLegalNumbering ? "decimal" : levelInfo?.numFmt || "decimal",
+      );
+      computedMarker = computedMarker.replace(placeholder, formatted);
+    }
+  }
+
+  listRendering.marker = computedMarker;
+};
+
+const enrichParagraphTextBoxes = (
+  paragraph: Paragraph,
+  paraXml: XmlElement,
+  styles: StyleMap | null,
+  theme: Theme | null,
+  numbering: NumberingMap | null,
+  rels: RelationshipMap | null,
+  media: Map<string, MediaFile> | null,
+): void => {
+  if (paragraph.content.length === 0) {
+    return;
+  }
+
+  const xmlChildren = getChildElements(paraXml);
+  let parsedIndex = 0;
+  let lastConsumedRun: Run | undefined;
+
+  for (const xmlChild of xmlChildren) {
+    if (getLocalName(xmlChild.name ?? "") !== "r") {
+      if (
+        parsedIndex < paragraph.content.length &&
+        paragraph.content[parsedIndex]?.type !== "run"
+      ) {
+        parsedIndex += 1;
+      }
+      continue;
+    }
+
+    const { textBoxDrawings, hasNonTextBoxContent } =
+      scanRunForTextBoxDrawings(xmlChild);
+
+    const parsedContent = paragraph.content[parsedIndex];
+    const parsedRun: Run | undefined =
+      parsedContent?.type === "run" ? parsedContent : undefined;
+    const targetRun =
+      parsedRun ?? (hasNonTextBoxContent ? lastConsumedRun : undefined);
+
+    for (const runEl of textBoxDrawings) {
+      const textBox = parseTextBox(runEl);
+      if (!textBox) {
+        continue;
+      }
+
+      const wsp = findDeep(runEl, "wps", "wsp");
+      if (wsp) {
+        const txbxContentEl = getTextBoxContentElement(wsp);
+        if (txbxContentEl) {
+          textBox.content = parseTextBoxContent(
+            txbxContentEl,
+            parseParagraph,
+            null,
+            styles,
+            theme,
+            numbering,
+            rels ?? undefined,
+            media ?? undefined,
+          );
+        }
+      }
+
+      const shape: Shape = {
+        type: "shape",
+        shapeType: "textBox",
+        size: textBox.size,
+        ...(textBox.position !== undefined
+          ? { position: textBox.position }
+          : {}),
+        ...(textBox.wrap !== undefined ? { wrap: textBox.wrap } : {}),
+        ...(textBox.fill !== undefined ? { fill: textBox.fill } : {}),
+        ...(textBox.outline !== undefined ? { outline: textBox.outline } : {}),
+        textBody: {
+          content: textBox.content,
+          ...(textBox.margins !== undefined
+            ? { margins: textBox.margins }
+            : {}),
+        },
+      };
+      if (textBox.id) {
+        shape.id = textBox.id;
+      }
+
+      const shapeContent: ShapeContent = { type: "shape", shape };
+
+      if (targetRun && hasNonTextBoxContent) {
+        targetRun.content.push(shapeContent);
+      } else {
+        const newRun: Run = { type: "run", content: [shapeContent] };
+        paragraph.content.splice(parsedIndex, 0, newRun);
+        lastConsumedRun = newRun;
+        parsedIndex += 1;
+      }
+    }
+
+    if (hasNonTextBoxContent && parsedRun) {
+      lastConsumedRun = parsedRun;
+      parsedIndex += 1;
+    }
+  }
+};
+
+type TextBoxRunScan = {
+  textBoxDrawings: XmlElement[];
+  hasNonTextBoxContent: boolean;
+};
+
+const scanRunForTextBoxDrawings = (xmlRun: XmlElement): TextBoxRunScan => {
+  const textBoxDrawings: XmlElement[] = [];
+  let hasNonTextBoxContent = false;
+
+  const visitDrawing = (drawingEl: XmlElement): void => {
+    if (isTextBoxDrawing(drawingEl)) {
+      textBoxDrawings.push(drawingEl);
+      return;
+    }
+    hasNonTextBoxContent = true;
+  };
+
+  for (const el of getChildElements(xmlRun)) {
+    const name = getLocalName(el.name ?? "");
+    if (name === "rPr") {
+      continue;
+    }
+    if (name === "drawing") {
+      visitDrawing(el);
+      continue;
+    }
+    if (name === "AlternateContent") {
+      const branches = getChildElements(el);
+      const choice = branches.find(
+        (branch) => getLocalName(branch.name ?? "") === "Choice",
+      );
+      const fallback = branches.find(
+        (branch) => getLocalName(branch.name ?? "") === "Fallback",
+      );
+      const tryBranch = (branch: XmlElement | undefined): boolean => {
+        if (!branch) {
+          return false;
+        }
+        let found = false;
+        for (const innerEl of getChildElements(branch)) {
+          if (getLocalName(innerEl.name ?? "") === "drawing") {
+            visitDrawing(innerEl);
+            found = true;
+          }
+        }
+        return found;
+      };
+      let foundInBranch = tryBranch(choice);
+      if (!foundInBranch) {
+        foundInBranch = tryBranch(fallback);
+      }
+      if (!foundInBranch) {
+        hasNonTextBoxContent = true;
+      }
+      continue;
+    }
+    hasNonTextBoxContent = true;
+  }
+
+  return { textBoxDrawings, hasNonTextBoxContent };
+};
+
+type ParseBlockContentState = {
+  listCounters: Map<number, number[]>;
+  abstractCounters: Map<number, number[]>;
+  options: ParseBlockContentOptions | undefined;
+};
+
+export const parseBlockContent = (
+  parent: XmlElement,
+  styles: StyleMap | null,
+  theme: Theme | null,
+  numbering: NumberingMap | null,
+  rels: RelationshipMap | null,
+  media: Map<string, MediaFile> | null,
+  options?: ParseBlockContentOptions,
+): (Paragraph | Table)[] =>
+  parseBlockContentWithState(parent, styles, theme, numbering, rels, media, {
+    listCounters: new Map(),
+    abstractCounters: new Map(),
+    options,
+  });
+
+const parseBlockContentWithState = (
+  parent: XmlElement,
+  styles: StyleMap | null,
+  theme: Theme | null,
+  numbering: NumberingMap | null,
+  rels: RelationshipMap | null,
+  media: Map<string, MediaFile> | null,
+  state: ParseBlockContentState,
+): (Paragraph | Table)[] => {
+  const content: (Paragraph | Table)[] = [];
+  const children = getChildElements(parent);
+
+  for (const child of children) {
+    const name = child.name ?? "";
+
+    if (name === "w:p" || name.endsWith(":p")) {
+      const paragraph = parseParagraph(
+        child,
+        styles,
+        theme,
+        numbering,
+        rels,
+        media,
+        state.options,
+      );
+      enrichParagraphTextBoxes(
+        paragraph,
+        child,
+        styles,
+        theme,
+        numbering,
+        rels,
+        media,
+      );
+      computeListMarker(
+        paragraph,
+        numbering,
+        state.listCounters,
+        state.abstractCounters,
+      );
+      content.push(paragraph);
+      continue;
+    }
+
+    if (name === "w:tbl" || name.endsWith(":tbl")) {
+      const table = parseTable(
+        child,
+        styles,
+        theme,
+        numbering,
+        rels,
+        media,
+        state.options,
+      );
+      content.push(table);
+      continue;
+    }
+
+    if (name === "w:sdt" || name.endsWith(":sdt")) {
+      const sdtContent = (child.elements ?? []).find(
+        (el: XmlElement) =>
+          el.type === "element" &&
+          (el.name === "w:sdtContent" || el.name?.endsWith(":sdtContent")),
+      );
+      if (sdtContent) {
+        content.push(
+          ...parseBlockContentWithState(
+            sdtContent,
+            styles,
+            theme,
+            numbering,
+            rels,
+            media,
+            state,
+          ),
+        );
+      }
+    }
+  }
+
+  return content;
+};
